+++ packages/orq-rc/src/models/components/evaluatorresponsetypescript.ts
@@ -32,8 +32,8 @@ export const EvaluatorResponseTypescript$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-08-24T14:41:02.470Z"),
-  updated: z.string().default("2026-08-24T14:41:02.470Z"),
+  created: z.string().default("2026-08-25T03:48:45.450Z"),
+  updated: z.string().default("2026-08-25T03:48:45.450Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   project_id: z.string().optional(),
   guardrail_config: z.any().optional(),