+++ packages/orq-rc/src/models/operations/updateannotationqueue.ts
@@ -195,7 +195,7 @@ export const UpdateAnnotationQueueResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-24T14:41:00.800Z",
+    "2026-08-25T03:48:43.842Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {